Next:ChatRoomMessagesCRUD

+++ FirstDjangoProject/base/views.py
@@ -1,5 +1,12 @@
 from django.shortcuts import render, redirect
-from .models import Room
+from django.http import HttpResponse
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
+from .models import Room, Topic
 from .forms import RoomForm
 
 """rooms = [
@@ -7,11 +14,61 @@ from .forms import RoomForm
     {'id':2, 'name':'Design with me'},
     {'id':3, 'name':'Frontend devs'},
 ]"""
+def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+    
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password') 
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'Invalid User')
+
+        user = authenticate(request, username=username, password=password)
+        
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username or Password does not exist')
+        
+    context = {'page': page}
+    return render(request, 'base/login_register.html', context)
+    
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    form = UserCreationForm()
+    context = {'form': form}
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occurred during registration')
+    
+    return render(request, 'base/login_register.html', context)
 
 # Create your views here.
 def home(request):
-    rooms = Room.objects.all()
-    context = {'rooms': rooms}
+    q = request.GET.get('q') if request.GET.get('q') != None else ''
+    
+    rooms = Room.objects.filter(Q(topic__name__icontains=q) | Q(name__icontains=q) | Q(description__icontains=q))
+    
+    topics = Topic.objects.all()
+    
+    room_count = rooms.count()
+    
+    context = {'rooms': rooms, 'topics': topics, 'room_count': room_count}
     return render(request, 'base/home.html', context)
 
 def room(request, pk):
@@ -19,8 +76,11 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
+    if request.user != room.host :
+        return HttpResponse('You are not allowed to be here')
     if request.method == 'POST':
         form = RoomForm(request.POST)
         if form.is_valid():
@@ -30,9 +90,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    
+    if request.user != room.host :
+        return HttpResponse('You are not allowed to be here')
     
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -43,8 +107,11 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host :
+        return HttpResponse('You are not allowed to be here')
     if request.method == 'POST':
         room.delete()
         return redirect('home')
